move action_handler to its own component
