+++ crates/astria-sequencer/src/accounts/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod action;
 pub(crate) mod component;
 pub(crate) mod query;
 mod state_ext;
