+++ crates/astria-sequencer/src/authority/mod.rs
@@ -1,4 +1,3 @@
-mod action;
 pub(crate) mod component;
 mod state_ext;
 pub(crate) mod storage;
