+++ crates/astria-sequencer/src/bridge/mod.rs
@@ -1,7 +1,3 @@
-mod bridge_lock_action;
-mod bridge_sudo_change_action;
-mod bridge_unlock_action;
-pub(crate) mod init_bridge_account_action;
 pub(crate) mod query;
 mod state_ext;
 pub(crate) mod storage;
