+++ crates/astria-sequencer/src/fees/mod.rs
@@ -37,7 +37,6 @@ use crate::{
     transaction::StateReadExt as _,
 };
 
-pub(crate) mod action;
 pub(crate) mod component;
 pub(crate) mod query;
 mod state_ext;
