+++ crates/astria-sequencer/src/ibc/mod.rs
@@ -1,8 +1,6 @@
 pub(crate) mod component;
 pub(crate) mod host_interface;
-pub(crate) mod ibc_relayer_change;
 pub(crate) mod ics20_transfer;
-pub(crate) mod ics20_withdrawal;
 mod state_ext;
 pub(crate) mod storage;
