+++ crates/astria-sequencer/src/lib.rs
@@ -1,4 +1,5 @@
 pub(crate) mod accounts;
+pub(crate) mod action_handler;
 pub(crate) mod address;
 pub(crate) mod app;
 pub(crate) mod assets;
@@ -15,7 +16,6 @@ pub(crate) mod ibc;
 mod mempool;
 pub(crate) mod metrics;
 pub(crate) mod proposal;
-pub(crate) mod rollup_data;
 mod sequencer;
 pub(crate) mod service;
 pub(crate) mod storage;
