--- crates/astria-sequencer/src/rollup_data/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod action;
