+++ crates/astria-sequencer/src/service/mempool/mod.rs
@@ -49,17 +49,12 @@ use tracing::{
 };
 
 use crate::{
-    accounts::StateReadExt as _,
-    address::StateReadExt as _,
-    app::ActionHandler as _,
-    mempool::{
+    accounts::StateReadExt as _, action_handler::ActionHandler as _, address::StateReadExt as _, mempool::{
         get_account_balances,
         InsertionError,
         Mempool as AppMempool,
         RemovalReason,
-    },
-    metrics::Metrics,
-    transaction,
+    }, metrics::Metrics, transaction
 };
 
 #[cfg(test)]
